Use shared-prefs to avoid onboarding screen that blocks the UI in expo dev client (#374)
This PR gets rid of the onboarding modal that pops up with
expo-dev-client apps after launch.

Since the modal is blocking JS looper, it prevents the app from
establishing devtools connection which may lead to timeouts and hence
prevent "app ready" events. As a consequence, the IDE fails to recognize
the app is properly launched.

Before, we've been using `disableOnboarding` flag passed as URL
parameter when launching the app. However, it appears like it's never
worked properly (see this fix here:
https://github.com/expo/expo/pull/29697). For the time being we decided
to use an alternative approach in which we write onboarding parameter
into shared props that are used to store the information whether the
popup should be displayed or not.