Add device setting to control privacy permissions (#423)
This PR adds a device setting that allows for resetting privacy
permissions.

The new setting helps with scenarios where you want to test flows that
rely on asking users for different types of permissions.

The current implementations supports resetting the following permissions
on iOS:
 - all
 - location
 - photos
 - contatcs
 - calendar

On Android due to the limitations of the method we currently use, only
resetting all permissions is supported.

On top of the above, the app may behave differently depending on the
system. On Android, the app process is always terminated when we reset
permission if at least one permission was granted prior to reset. On iOS
the app never restarts.

An important consideration here is that we may want to restart the app
regardless of the operating system, as the behavior where the app keeps
running may lead to some unpredictable behavior. This may also end up
getting the app in a state that is not expected by the developers.
Typically for end-users on Android and iOS it is not possible to reset
the privacy settings – you can change the setting to disallow or grant
the permission later on, but you can never get it back to the state when
it's unknown unless you uninstall the app. So this new option we are
adding here may get the app into a state that will never be possible to
replicate in real-life and therefore there is no need to handle such
state for the developer.

The biggest downside of the current implementation is that we use the
restart flow for the case when the permission is reset on Android. This
leads to a longer than necessary time to interact with the app, as the
current restart flow will also reboot the emulator, potentially trigger
build and also restart metro. I plan to address this in a follow up PR,
where we will add a new quicker flow for handling restarts that doesn't
reset the emulator and metro.

Here's how the new dropdown menu look for iOS:
<img width="535" alt="image"
src="https://github.com/software-mansion/react-native-ide/assets/726445/fafb68e1-251c-44d0-af1d-e8b444149652">

Test plan:
We're adding new location screen to expo-router example and extending
that example to ask for location permission.
This has been tested on this new example with two scenarios: 1) get
location info, allow, reset, get again 2) get location info, disallow,
reset permission, get again. In both scenarios it is expected to get the
permission popup. On Android the app will always restart when resetting
the permission.