Suppress expected errors for certain CDP commands (#599)
Previously, we added `catch` statements to avoid certain, expected
errors from being reported in the log output. However, this turns out to
not work, as it seems the catch with no handler method doesn't seem to
catch the exception.

This PR adds a no-op handler for catch statement which makes the error
no longer appear in the extension output.