Redirect logs/uncaught errors from webview to the main extension logger (#621)
This PR adds a mechanism for redirecting logs and unhandled error from
webview to the main extension.

The reason for this change is that we can see potential issues and logs
in a unified log output that we create for the extension context. This
way we provide better visibility of errors that happen in the webview.
Without this change, we'd always need to check both the extension log
output, and the vscode devtools panel. With this change, having
everything in a single output is specifically helpful when receiving
logs from the users, as we no longer need to ask them to check devtools
console and can focus on a single log file they need to check or send as
a part of the report.

### How Has This Been Tested: 
Add log statement and unhandled error to the main webview App.tsx. See
that this gets included in the main log output.