Redo the diagnostics logic (#615)
This PR aims to address several issues related to diagnostics / setup
checks that we do in the IDE:
1. Occasionally I'd be getting a crash in webview panel due to
diagnostics events coming too early and triggering a code-path that
would throw uncaught exception. In that scenario IDE panel would be just
empty w/o way of recovering
2. Diagnostics dialog would stay up all the time w/o ways of closing it
in certain scenarios. This was annoying as it'd stay up despite the fact
the whole setup was successful and the app would boot up normally.
3. Pod install would report error in some cases, but there was no way of
recovering from those if the build was ok and cached.

What we are changing to address those:
1. We are remocing diagnostics dialog entirely. You can still preview
diagnostics in the separate modal.
2. We are restructuring the front-end code of diagnostics. We simplify
the DepsManager protocol and it now provides a way to re-run checks, and
listener for changes. The pervious crashes were due to the fact the
updates were stored in state while we'd sometimes rely on state being
updated immediately while it isn't the case with React (state only
updates with the next render)
3. We are restructuring implementation of DepsManager on the extension
side: it no longer uses string identifier for checks. Also we are
renaming "isInstalled" to "check<Something>" to indicate that we indeed
perform the check when the method is called rather than return some
cached value.
4. We are renaming a number of methods inside DepsManager to better
reflect what they do (i.e. we had isInstalledPods and
isInstalledCocoapods before, where one was checking whether the command
exists and the other was checking whether we have the pods folder with
dependencies installed)
5. We are updating the build logic such that it ignores cached builds
when pod installation is required. This of course handles the case when
custom build option is set, in which case we ignore pods checks (they
are treated as optional).
6. We move platform checks for filtering diagnostics entries from
webview to extension (they fit much better there anyways)

### How Has This Been Tested: 
1. Test diagnostics modal on one of test-apps project
2. Full acceptance test suite on expo-router test app + storybook
3. Test app workflow on yarn project (reanimated 4)