Use defaults command to update scheme approval settings (#622)
This PR changes the way we update scheme approval list for launching
apps via deep links.

Previously, we'd manually modify plist files under the simulator
"filesystem". This wouldn't work in some cases, as the content of the
file could be cached by some of the system services. Modifying it using
the defaults command seem to resolve the issue and always results in the
simulator reflecting the made changes.

For some reason the previous approach would work most of the time, as
we'd typically boot the device prior to launching. However, under
certain cirmustances (that are difficult to replicate), it would use a
booted version that already had the plist file loaded. As a consequence,
the app would get stuck on the "Waiting for app to load" message as the
app process would get stuck on the system dialog asking whether the link
can be opened with the installed app.

Fixes #544

### How Has This Been Tested: 
1. Create new iOS simulator - launch expo dev client app and expo go app
2. Use `defaults read com.apple.launchservices.schemeapproval` to read
the updated content.