Only consider pods check when building for iOS (#628)
This fixes a bug introduced in #615 that was causing Android rebuilds
when pod check was failing.

In #615 we added a cocoapods check before starting a native build, which
was then used as an additional indicator of when we should rebuild (when
check was failing we'd clean build cache). This check however only makes
sense for iOS builds. And in the scenario when pods aren't yet installed
and you try to build for Android, it was causing native rebuilds every
time.

The change here is to make the check only when building for iOS.

We're also adding some additional logs that should indicate the native
build reason better in the logs.

### How Has This Been Tested: 
1. Checkout clean copy of expo-51-dev-client
2. Open it for Android at first
3. Restart IDE card and see it no longer performs native build. Before
this change, every time it'd start a new gradle build (which was
relatively fast because it was an incremental build) but still should be
avoided.